# services/backend/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "AdminUserListResponse": {
        "additionalProperties": false,
        "properties": {
          "admin_users": {
            "items": {
              "$ref": "#/components/schemas/AdminUserView"
            },
            "title": "Admin Users",
            "type": "array"
          }
        },
        "required": [
          "admin_users"
        ],
        "title": "AdminUserListResponse",
        "type": "object"
      },
      "AdminUserView": {
        "additionalProperties": false,
        "description": "Never `password_hash`, and never the lockout counters.\n\n`extra=\"forbid\"` is what makes that structural rather than a habit: adding the hash to\nthe command result without adding it here fails at serialisation, and\n`test_no_admin_user_response_carries_a_credential` fails if either changes.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "phone_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "record_version": {
            "title": "Record Version",
            "type": "integer"
          },
          "role_codes": {
            "items": {
              "type": "string"
            },
            "title": "Role Codes",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "full_name",
          "email",
          "phone_number",
          "status",
          "role_codes",
          "record_version"
        ],
        "title": "AdminUserView",
        "type": "object"
      },
...
      "CreateAdminUserRequest": {
        "additionalProperties": false,
        "description": "No `status` field: a new account's state is the command's decision, not the\ncaller's, and accepting it would let a creation mint a `deactivated` account or one in\n`recovery_required` that nothing can recover.",
        "properties": {
          "email": {
            "anyOf": [
              {
                "maxLength": 254,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "full_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "phone_number": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Number"
          },
          "role_codes": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Role Codes",
            "type": "array"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "full_name",
          "password",
          "role_codes"
        ],
        "title": "CreateAdminUserRequest",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "RoleListResponse": {
        "additionalProperties": false,
        "properties": {
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleView"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "roles"
        ],
        "title": "RoleListResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OperationsToken": {
        "description": "Operations-only health access token.",
        "in": "header",
        "name": "X-Operations-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned operational API. Financial workflow modules are intentionally not part of the M1 runtime foundation.",
    "summary": "Manual-first settlement platform backend",
    "title": "Gold Trade Settlement Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin-users": {
      "get": {
        "operationId": "listAdminUsers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          }
        },
        "summary": "List staff accounts.",
        "tags": [
          "admin-users"
        ]
      },
      "post": {
        "description": "`Idempotency-Key` is required **and used**.\n\nDoc 05:868 requires the header. `12_Security_RBAC_Audit.md` §12 additionally requires\nthe record to be resolved and its completion persisted, which the four trader decision\nroutes do not do — they require the header and discard it. This one claims through\n`IdempotencyResolver`, so a retried creation returns the first account instead of a\nsecond person with the same name.",
        "operationId": "createAdminUser",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request violates a business rule."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such account."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The idempotency key was reused with a different body."
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The If-Match value is stale."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "If-Match or Idempotency-Key is missing."
          }
        },
        "summary": "Create a staff account.",
        "tags": [
          "admin-users"
        ]
      }
    },
    "/api/v1/admin-users/{admin_user_id}": {
      "get": {
        "operationId": "getAdminUser",
        "parameters": [
          {
            "in": "path",
            "name": "admin_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Admin User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          }
        },
        "summary": "Read one staff account.",
        "tags": [
          "admin-users"
        ]
      },
      "patch": {
        "description": "`If-Match` required, per doc 05:870.",
        "operationId": "updateAdminUser",
        "parameters": [
          {
            "in": "path",
            "name": "admin_user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Admin User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AmendAdminUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request violates a business rule."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such account."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The idempotency key was reused with a different body."
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The If-Match value is stale."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "If-Match or Idempotency-Key is missing."
          }
        },
        "summary": "Amend a staff account's contact details.",
        "tags": [
          "admin-users"
        ]
      }
    },
...
    "/api/v1/roles": {
      "get": {
        "operationId": "listRoles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such role."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          }
        },
        "summary": "List roles and the permissions each carries.",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{role_id}": {
      "get": {
        "operationId": "getRole",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such role."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          }
        },
        "summary": "Read one role. Publishes the ETag the update must echo.",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{role_id}/permissions": {
      "put": {
        "description": "Both preconditions, and they answer different questions.\n\n`If-Match` stops this landing on a permission set somebody else has changed.\n`X-Recent-Auth` proves the person holding the session was present a moment ago. Neither\nsubstitutes for the other: a fresh step-up does not make a stale write safe, and an\nup-to-date digest says nothing about who is holding the browser.\n\nThe step-up must have been issued for **this role** — purpose `role.permissions.update`,\nresource type `role`, resource id this one. A context obtained to edit `accountant`\npresented here is refused as `WRONG_RESOURCE`, which is the binding the whole step-up\ndesign exists for.",
        "operationId": "updateRolePermissions",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "X-Recent-Auth",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recent-Auth"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request violates a business rule."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid session."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Permission denied."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No such role."
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The If-Match digest is stale."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request failed canonical validation."
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "If-Match or X-Recent-Auth is missing."
          },
          "440": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recent authentication is required, or the one presented does not authorise this change."
          }
        },
        "summary": "Replace a role's permission set. Requires recent authentication.",
        "tags": [
          "roles"
        ]
      }
    },
...
  }
}